Basic UI Layout - closes #1
