+++ resources/views/home.blade.php
@@ -29,209 +29,10 @@
             <a href="#" class="hover:text-green-600 transition">Become a Pro</a>
         </div>
         <div class="flex space-x-4">
-            <button onclick="toggleModal('loginModal')" class="px-5 py-2 text-green-600 font-semibold hover:bg-green-50 rounded-lg transition">Login</button>
-            <button onclick="toggleModal('registerModal')" class="px-5 py-2 bg-green-500 text-white font-semibold rounded-lg hover:bg-green-600 shadow-md transition">Sign Up</button>
+            <a href="/login" class="px-5 py-2 text-green-600 font-semibold hover:bg-green-50 rounded-lg transition">Login</a>
+            <a href="/signup" class="px-5 py-2 bg-green-500 text-white font-semibold rounded-lg hover:bg-green-600 shadow-md transition">Sign Up</a>
         </div>
     </nav>
-
-    <!-- Login Modal -->
-    <div id="loginModal" class="fixed inset-0 z-[100] hidden overflow-y-auto" aria-labelledby="modal-title" role="dialog" aria-modal="true">
-        <div class="flex items-center justify-center min-h-screen px-4 pt-4 pb-20 text-center sm:block sm:p-0">
-            <!-- Background overlay -->
-            <div class="fixed inset-0 transition-opacity bg-gray-900 bg-opacity-75" aria-hidden="true" onclick="toggleModal('loginModal')"></div>
-
-            <span class="hidden sm:inline-block sm:align-middle sm:h-screen" aria-hidden="true">&#8203;</span>
-
-            <div class="inline-block align-bottom bg-white rounded-2xl text-left overflow-hidden shadow-2xl transform transition-all sm:my-8 sm:align-middle sm:max-w-md sm:w-full border border-green-100">
-                <div class="bg-white px-8 pt-10 pb-8">
-                    <div class="text-center mb-8">
-                        <div id="signupSuccessMsg" class="hidden mb-6 p-4 bg-green-100 text-green-700 rounded-xl text-sm font-bold animate-bounce">
-                            <i class="fas fa-check-circle mr-2"></i> Registration Successful! Please Login.
-                        </div>
-                        <div class="w-16 h-16 bg-green-100 text-green-600 rounded-full flex items-center justify-center mx-auto mb-4">
-                            <i class="fas fa-user-lock text-2xl"></i>
-                        </div>
-                        <h3 class="text-2xl font-bold text-gray-900" id="modal-title">Welcome Back</h3>
-                        <p class="text-gray-500 mt-2">Login to manage your bookings</p>
-                    </div>
-
-                    <form action="/" method="GET" class="space-y-6">
-                        <div>
-                            <label class="block text-sm font-semibold text-gray-700 mb-2">Email Address</label>
-                            <div class="relative">
-                                <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                    <i class="fas fa-envelope"></i>
-                                </span>
-                                <input type="email" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="name@example.com">
-                            </div>
-                        </div>
-
-                        <div>
-                            <div class="flex justify-between mb-2">
-                                <label class="block text-sm font-semibold text-gray-700">Password</label>
-                                <a href="#" class="text-sm font-medium text-green-600 hover:text-green-700">Forgot password?</a>
-                            </div>
-                            <div class="relative">
-                                <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                    <i class="fas fa-lock"></i>
-                                </span>
-                                <input type="password" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="••••••••">
-                            </div>
-                        </div>
-
-                        <div class="flex items-center">
-                            <input type="checkbox" class="h-4 w-4 text-green-600 focus:ring-green-500 border-gray-300 rounded">
-                            <label class="ml-2 block text-sm text-gray-700">Remember me</label>
-                        </div>
-
-                        <button type="submit" class="w-full py-4 bg-green-500 text-white font-bold rounded-xl hover:bg-green-600 shadow-lg shadow-green-200 transition-all transform hover:-translate-y-0.5">
-                            Sign In
-                        </button>
-                    </form>
-
-                    <div class="mt-8 text-center border-t border-gray-100 pt-6">
-                        <p class="text-gray-600">Don't have an account? 
-                            <button onclick="toggleModal('loginModal'); toggleModal('registerModal')" class="font-bold text-green-600 hover:text-green-700">Create Account</button>
-                        </p>
-                    </div>
-                </div>
-                <button onclick="toggleModal('loginModal')" class="absolute top-4 right-4 text-gray-400 hover:text-gray-600 transition">
-                    <i class="fas fa-times text-xl"></i>
-                </button>
-            </div>
-        </div>
-    </div>
-
-    <!-- Register Modal -->
-    <div id="registerModal" class="fixed inset-0 z-[100] hidden overflow-y-auto" aria-labelledby="modal-title" role="dialog" aria-modal="true">
-        <div class="flex items-center justify-center min-h-screen px-4 pt-4 pb-20 text-center sm:block sm:p-0">
-            <div class="fixed inset-0 transition-opacity bg-gray-900 bg-opacity-75" aria-hidden="true" onclick="toggleModal('registerModal')"></div>
-            <span class="hidden sm:inline-block sm:align-middle sm:h-screen" aria-hidden="true">&#8203;</span>
-            <div class="inline-block align-bottom bg-white rounded-2xl text-left overflow-hidden shadow-2xl transform transition-all sm:my-8 sm:align-middle sm:max-w-md sm:w-full border border-green-100">
-                <div class="bg-white px-8 pt-10 pb-8">
-                    <div class="text-center mb-8">
-                        <div class="w-16 h-16 bg-green-100 text-green-600 rounded-full flex items-center justify-center mx-auto mb-4">
-                            <i class="fas fa-user-plus text-2xl"></i>
-                        </div>
-                        <h3 class="text-2xl font-bold text-gray-900">Create Account</h3>
-                        <p class="text-gray-500 mt-2">Join Serviqo today</p>
-                    </div>
-
-                    <form id="registerForm" action="/" method="GET" class="space-y-4" onsubmit="return validatePasswords()">
-                        <input type="hidden" name="signup" value="success">
-                        <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                                <label class="block text-sm font-semibold text-gray-700 mb-1">Full Name</label>
-                                <div class="relative">
-                                    <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                        <i class="fas fa-user"></i>
-                                    </span>
-                                    <input type="text" name="name" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="John Doe">
-                                </div>
-                            </div>
-                            <div>
-                                <label class="block text-sm font-semibold text-gray-700 mb-1">Email Address</label>
-                                <div class="relative">
-                                    <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                        <i class="fas fa-envelope"></i>
-                                    </span>
-                                    <input type="email" name="email" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="name@example.com">
-                                </div>
-                            </div>
-                        </div>
-
-                        <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                                <label class="block text-sm font-semibold text-gray-700 mb-1">Phone Number</label>
-                                <div class="relative">
-                                    <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                        <i class="fas fa-phone"></i>
-                                    </span>
-                                    <input type="tel" name="phone" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="+880 1XXX-XXXXXX">
-                                </div>
-                            </div>
-                            <div>
-                                <label class="block text-sm font-semibold text-gray-700 mb-1">Date of Birth</label>
-                                <div class="relative">
-                                    <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                        <i class="fas fa-calendar-alt"></i>
-                                    </span>
-                                    <input type="date" name="dob" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition text-gray-700">
-                                </div>
-                            </div>
-                        </div>
-
-                        <div>
-                            <label class="block text-sm font-semibold text-gray-700 mb-1">Division</label>
-                            <div class="relative">
-                                <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400 pointer-events-none">
-                                    <i class="fas fa-map-marker-alt"></i>
-                                </span>
-                                <select name="division" required class="block w-full pl-10 pr-10 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition appearance-none bg-white text-gray-700">
-                                    <option value="" disabled selected>Select Division</option>
-                                    <option value="Dhaka">Dhaka</option>
-                                    <option value="Chittagong">Chittagong</option>
-                                    <option value="Sylhet">Sylhet</option>
-                                    <option value="Barisal">Barisal</option>
-                                    <option value="Rangpur">Rangpur</option>
-                                    <option value="Rajshahi">Rajshahi</option>
-                                    <option value="Khulna">Khulna</option>
-                                </select>
-                                <div class="absolute inset-y-0 right-0 flex items-center pr-3 pointer-events-none text-gray-400">
-                                    <i class="fas fa-chevron-down text-xs"></i>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div>
-                            <label class="block text-sm font-semibold text-gray-700 mb-1">Full Address</label>
-                            <div class="relative">
-                                <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                    <i class="fas fa-home"></i>
-                                </span>
-                                <textarea name="address" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" rows="2" placeholder="Street, Apartment, Zip Code"></textarea>
-                            </div>
-                        </div>
-
-                        <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                                <label class="block text-sm font-semibold text-gray-700 mb-1">Password</label>
-                                <div class="relative">
-                                    <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                        <i class="fas fa-lock"></i>
-                                    </span>
-                                    <input type="password" id="regPassword" name="password" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="••••••••">
-                                </div>
-                            </div>
-                            <div>
-                                <label class="block text-sm font-semibold text-gray-700 mb-1">Confirm Password</label>
-                                <div class="relative">
-                                    <span class="absolute inset-y-0 left-0 pl-3 flex items-center text-gray-400">
-                                        <i class="fas fa-lock"></i>
-                                    </span>
-                                    <input type="password" id="regConfirmPassword" name="password_confirmation" required class="block w-full pl-10 pr-3 py-3 border border-gray-200 rounded-xl focus:ring-2 focus:ring-green-500 focus:border-transparent outline-none transition" placeholder="••••••••">
-                                </div>
-                            </div>
-                        </div>
-                        <p id="passwordError" class="text-red-500 text-sm hidden">Passwords do not match!</p>
-
-                        <button type="submit" class="w-full py-4 bg-green-500 text-white font-bold rounded-xl hover:bg-green-600 shadow-lg shadow-green-200 transition-all transform hover:-translate-y-0.5 mt-2">
-                            Create Account
-                        </button>
-                    </form>
-
-                    <div class="mt-8 text-center border-t border-gray-100 pt-6">
-                        <p class="text-gray-600">Already have an account? 
-                            <button onclick="toggleModal('registerModal'); toggleModal('loginModal')" class="font-bold text-green-600 hover:text-green-700">Login Now</button>
-                        </p>
-                    </div>
-                </div>
-                <button onclick="toggleModal('registerModal')" class="absolute top-4 right-4 text-gray-400 hover:text-gray-600 transition">
-                    <i class="fas fa-times text-xl"></i>
-                </button>
-            </div>
-        </div>
-    </div>
 
     <!-- Hero Section -->
     <header class="relative bg-gradient-to-br from-green-50 to-white py-20 lg:py-32 overflow-hidden">
@@ -538,51 +339,5 @@
             </div>
         </div>
     </footer>
-
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            console.log('Page loaded, checking for signup status...');
-            const urlParams = new URLSearchParams(window.location.search);
-            if (urlParams.get('signup') === 'success') {
-                console.log('Signup success detected! Opening login modal...');
-                const successMsg = document.getElementById('signupSuccessMsg');
-                if (successMsg) {
-                    successMsg.classList.remove('hidden');
-                }
-                
-                setTimeout(() => {
-                    toggleModal('loginModal');
-                }, 500); // Slightly longer delay for safety
-
-                // Clean up the URL
-                const newUrl = window.location.pathname;
-                window.history.replaceState({}, document.title, newUrl);
-            }
-        });
-
-        function toggleModal(modalId) {
-            const modal = document.getElementById(modalId);
-            if (modal.classList.contains('hidden')) {
-                modal.classList.remove('hidden');
-                document.body.style.overflow = 'hidden';
-            } else {
-                modal.classList.add('hidden');
-                document.body.style.overflow = 'auto';
-            }
-        }
-
-        function validatePasswords() {
-            const pass = document.getElementById('regPassword').value;
-            const confirm = document.getElementById('regConfirmPassword').value;
-            const errorElement = document.getElementById('passwordError');
-
-            if (pass !== confirm) {
-                errorElement.classList.remove('hidden');
-                return false;
-            }
-            errorElement.classList.add('hidden');
-            return true;
-        }
-    </script>
 </body>
 </html>
